:rotating_light: merge all notion-related GET API into one to stay bellow vercel limit of serverless functions
