--- api/notion/test.js
@@ -1,152 +1,0 @@
-const test = [
-  {
-    object: 'page',
-    id: '9d7c520a-87d0-4777-beae-4e969605f7b1',
-    created_time: '2022-08-22T12:20:00.000Z',
-    last_edited_time: '2022-08-22T12:20:00.000Z',
-    created_by: { object: 'user', id: 'a229569e-897f-4af4-888c-cf78758fd7c0' },
-    last_edited_by: {
-      object: 'user',
-      id: 'a229569e-897f-4af4-888c-cf78758fd7c0',
-    },
-    cover: null,
-    icon: null,
-    parent: {
-      type: 'database_id',
-      database_id: '80dbe87a-5205-484e-8f14-b21614d54406',
-    },
-    archived: false,
-    properties: {
-      'quantité ingredient /pers': [Object],
-      Ingredients: [Object],
-      'Id Ingredients': [Object],
-      Nom: [Object],
-    },
-    url: 'https://www.notion.so/Spaghetti-aux-courgettes-9d7c520a87d04777beae4e969605f7b1',
-  },
-  {
-    object: 'page',
-    id: '15c307d7-7c4d-4b3f-ac07-1b00fddc38c2',
-    created_time: '2022-08-22T11:52:00.000Z',
-    last_edited_time: '2022-08-22T11:52:00.000Z',
-    created_by: { object: 'user', id: 'a229569e-897f-4af4-888c-cf78758fd7c0' },
-    last_edited_by: {
-      object: 'user',
-      id: 'a229569e-897f-4af4-888c-cf78758fd7c0',
-    },
-    cover: null,
-    icon: null,
-    parent: {
-      type: 'database_id',
-      database_id: '80dbe87a-5205-484e-8f14-b21614d54406',
-    },
-    archived: false,
-    properties: {
-      'quantité ingredient /pers': [Object],
-      Ingredients: [Object],
-      'Id Ingredients': [Object],
-      Nom: [Object],
-    },
-    url: 'https://www.notion.so/15c307d77c4d4b3fac071b00fddc38c2',
-  },
-  {
-    object: 'page',
-    id: '109d4888-c0fd-4387-8dc9-2aef4a8042eb',
-    created_time: '2022-08-21T20:53:00.000Z',
-    last_edited_time: '2022-08-21T20:53:00.000Z',
-    created_by: { object: 'user', id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d' },
-    last_edited_by: {
-      object: 'user',
-      id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d',
-    },
-    cover: null,
-    icon: null,
-    parent: {
-      type: 'database_id',
-      database_id: '80dbe87a-5205-484e-8f14-b21614d54406',
-    },
-    archived: false,
-    properties: {
-      'quantité ingredient /pers': [Object],
-      Ingredients: [Object],
-      'Id Ingredients': [Object],
-      Nom: [Object],
-    },
-    url: 'https://www.notion.so/P-tes-ratatouille-109d4888c0fd43878dc92aef4a8042eb',
-  },
-  {
-    object: 'page',
-    id: 'a1977531-4239-40d5-929c-20bd4551d09c',
-    created_time: '2022-08-07T17:45:00.000Z',
-    last_edited_time: '2022-08-07T17:45:00.000Z',
-    created_by: { object: 'user', id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d' },
-    last_edited_by: {
-      object: 'user',
-      id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d',
-    },
-    cover: null,
-    icon: null,
-    parent: {
-      type: 'database_id',
-      database_id: '80dbe87a-5205-484e-8f14-b21614d54406',
-    },
-    archived: false,
-    properties: {
-      'quantité ingredient /pers': [Object],
-      Ingredients: [Object],
-      'Id Ingredients': [Object],
-      Nom: [Object],
-    },
-    url: 'https://www.notion.so/bolo-a1977531423940d5929c20bd4551d09c',
-  },
-  {
-    object: 'page',
-    id: '008b4154-75b1-411e-a946-1b1b36d806ef',
-    created_time: '2022-08-07T17:37:00.000Z',
-    last_edited_time: '2022-08-07T17:46:00.000Z',
-    created_by: { object: 'user', id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d' },
-    last_edited_by: {
-      object: 'user',
-      id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d',
-    },
-    cover: null,
-    icon: null,
-    parent: {
-      type: 'database_id',
-      database_id: '80dbe87a-5205-484e-8f14-b21614d54406',
-    },
-    archived: false,
-    properties: {
-      'quantité ingredient /pers': [Object],
-      Ingredients: [Object],
-      'Id Ingredients': [Object],
-      Nom: [Object],
-    },
-    url: 'https://www.notion.so/Quiche-008b415475b1411ea9461b1b36d806ef',
-  },
-  {
-    object: 'page',
-    id: '5c4693f7-f4a2-48a6-a3e7-6e132ef2e851',
-    created_time: '2022-07-10T12:39:00.000Z',
-    last_edited_time: '2022-08-07T17:46:00.000Z',
-    created_by: { object: 'user', id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d' },
-    last_edited_by: {
-      object: 'user',
-      id: 'e4fa2538-1bde-4c93-8455-74a70fb08a5d',
-    },
-    cover: null,
-    icon: null,
-    parent: {
-      type: 'database_id',
-      database_id: '80dbe87a-5205-484e-8f14-b21614d54406',
-    },
-    archived: false,
-    properties: {
-      'quantité ingredient /pers': [Object],
-      Ingredients: [Object],
-      'Id Ingredients': [Object],
-      Nom: [Object],
-    },
-    url: 'https://www.notion.so/P-tes-carbo-5c4693f7f4a248a6a3e76e132ef2e851',
-  },
-];
