+++ hooks/useIngredients.ts
@@ -1,8 +1,8 @@
 import useSWR from 'swr';
-import { fetcher } from '../../helpers/fetcher';
+import { fetcher } from '../helpers/fetcher';
 
 export const useIngredients = () => {
-  const { data, error, mutate } = useSWR('/api/ingredients', fetcher);
+  const { data, error, mutate } = useSWR('/api/data/ingredients', fetcher);
   return {
     ingredients: data?.ingredients,
     isLoading: !error && !data,
