+++ hooks/useMeals.ts
@@ -1,13 +1,13 @@
 import { useState } from 'react';
 import useSWR from 'swr';
-import { fetcher } from '../../helpers/fetcher';
-import { useMealsProperties } from '../../types';
+import { fetcher } from '../helpers/fetcher';
+import { useMealsProperties } from '../types';
 
 export const useMeals = ({ firstDate, lastDate }: useMealsProperties) => {
   const { data, error } = useSWR(
-    `/api/meals?firstDate=${firstDate.toISOString().split('T')[0]}&lastDate=${
-      lastDate.toISOString().split('T')[0]
-    }`,
+    `/api/data/meals?firstDate=${
+      firstDate.toISOString().split('T')[0]
+    }&lastDate=${lastDate.toISOString().split('T')[0]}`,
     fetcher
   );
