+++ hooks/useReciepes.ts
@@ -1,8 +1,8 @@
 import useSWR from 'swr';
-import { fetcher } from '../../helpers/fetcher';
+import { fetcher } from '../helpers/fetcher';
 
 export const useReciepes = () => {
-  const { data, error } = useSWR('/api/reciepes', fetcher);
+  const { data, error } = useSWR('/api/data/reciepes', fetcher);
   return {
     reciepes: data?.reciepes,
     isLoading: !error && !data,
